requirements/staging.txt
-r base.txt

django-toolbelt==0.0.1
gunicorn==19.3.0
psycopg2==2.5.2
static3==0.5.1
dj-database-url==0.3.0
dj-static==0.0.6